replace instances of String with c-string

+++ source/libc/cstdio.cpp
@@ -342,6 +342,6 @@ void fstat(int32_t fd, FileStat* buffer) { stat(fd, (void*)buffer); }
 Array<File*> readdir(const char* filename)
 {
 	Array<File*> files;
-	asm("int $0x80" ::"a"(SYSCALL_READDIR), "b"(filename), "c"(&files));
+	asm("int $0x80" ::"a"(SYSCALL_READDIR), "b"(filename), "c"((int32_t)&files));
 	return files;
 }
